+++ src/test/java/performanceTest/Neo4J_performanceTest.java
@@ -13,50 +13,49 @@ import queries.QueryInterface;
  *
  * @author OpieOP
  */
-public class Neo4J_performanceTest  {
-    
-    private final String[] citySet = {"'Sar-e Pul'","'Szczecin'","'New York City'","'London'", "'Copenhagen'"};
-    private final String[] titleSet = {"Charlotte Brontë and Her Circle", "I Married a Ranger, A Short Method Of Prayer", "In the Track of the Bookworm", "Much Ado About Something","Bimbi, Stories for Children"};
-    private final String[] authorSet = {"Abbott, Edwin Abbott", "Karasowski, Maurycy" , "Taylor, Robert Bruce", "Sherman, Frederic, Mrs.", "Terry, Dorothy"};
-    private final float[][] coordinateSet = {{51f,0f},{36.2154f, 65.9325f},{53.4289f,14.553f},{40.7143f,-74.006f},{55.6759f,12.5655f}};
+public class Neo4J_performanceTest {
+
+    private final String[] citySet = {"'Sar-e Pul'", "'Szczecin'", "'New York City'", "'London'", "'Copenhagen'"};
+    private final String[] titleSet = {"Charlotte Brontë and Her Circle", "I Married a Ranger, A Short Method Of Prayer", "In the Track of the Bookworm", "Much Ado About Something", "Bimbi, Stories for Children"};
+    private final String[] authorSet = {"Abbott, Edwin Abbott", "Karasowski, Maurycy", "Taylor, Robert Bruce", "Sherman, Frederic, Mrs.", "Terry, Dorothy"};
+    private final float[][] coordinateSet = {{51f, 0f}, {36.2154f, 65.9325f}, {53.4289f, 14.553f}, {40.7143f, -74.006f}, {55.6759f, 12.5655f}};
     private static long start;
     private static long end;
-    
-    public ArrayList<Long> Neo4JPerformance() {
-        ArrayList<Long> results = new ArrayList();
+
+    public ArrayList<Double> Neo4JPerformance() {
+        ArrayList<Double> results = new ArrayList();
         start = System.currentTimeMillis();
         for (int i = 0; i < 5; i++) {
             getMentioningBooksWithAuthorTest(citySet[i]);
         }
         end = System.currentTimeMillis();
-        results.add((end-start)/1000L);
-        
+        results.add((double) (end - start) / 1000.00);
+
         start = System.currentTimeMillis();
         for (int i = 0; i < 5; i++) {
             mentionedCitiesTest(titleSet[i]);
         }
         end = System.currentTimeMillis();
-        results.add((end-start)/1000L);
-        
+        results.add((double) (end - start) / 1000.00);
+
         start = System.currentTimeMillis();
         for (int i = 0; i < 5; i++) {
             getBooksAndCitiesTest(authorSet[i]);
         }
         end = System.currentTimeMillis();
-        results.add((end-start)/1000L);
-        
+        results.add((double) (end - start) / 1000.00);
+
         start = System.currentTimeMillis();
         for (int i = 0; i < 5; i++) {
-            getBooksMentioningRange(coordinateSet[i][0],coordinateSet[i][1], 15000);
+            getBooksMentioningRange(coordinateSet[i][0], coordinateSet[i][1], 15000);
         }
         end = System.currentTimeMillis();
-        results.add((end-start)/1000L);
-        
+        results.add((double) (end - start) / 1000.00);
+
         return results;
-    
-    
+
     }
-    
+
     @Before
     public void setUp() {
     }
@@ -66,7 +65,6 @@ public class Neo4J_performanceTest  {
     //
     // @Test
     // public void hello() {}
-
     public static void getMentioningBooksWithAuthorTest(String cityname) {
         QueryInterface.getMentioningBooksWithAuthors(cityname, QueryInterface.DBChoice.DB_NEO4J);
     }
